pyproject.toml
[tool.poetry]
name = "agentic-kg-explorer"
version = "0.1.0"
description = "Agentic AI Knowledge Graph Explorer - LangGraph + Neo4j"
authors = ["Youngha"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.3.0"
langchain-openai = "^0.2.0"
langchain-anthropic = "^0.2.0"
langgraph = "^0.2.0"
neo4j = "^5.25.0"
python-dotenv = "^1.0.0"
pydantic = "^2.9.0"
pydantic-settings = "^2.5.0"
fastapi = "^0.115.0"
uvicorn = "^0.31.0"
httpx = "^0.27.0"
faker = "^40.1.2"
dotenv = "^0.9.9"
google-genai = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
black = "^24.8.0"
ruff = "^0.6.0"
mypy = "^1.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
